LayoutTests/media/modern-media-controls/media-controller/media-controller-scale-factor-expected.txt

Testing that we apply a CSS zoom to ensure controls are always 50 device points tall on iOS.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS shadowRoot.querySelector('.media-controls').style.width is "800px"
PASS shadowRoot.querySelector('.media-controls').style.height is "600px"
PASS shadowRoot.querySelector('.media-controls').style.zoom is "0.5"
PASS shadowRoot.querySelector('.media-controls').style.top is "auto"

PASS successfullyParsed is true

TEST COMPLETE
